Start work on client chat tab
Game assemblies are referenced now so I'm expecting TravisCI to flip out until I can figure out how to ignore UI-related code

The tab contents are placeholder for the most part. The message box and send button are off a little but everything else seems to be in place.
Also added an English keyed translation file so translation is supported from the get-go.