Event-ify server connection events and genericise Send method
Connection events (established/closed) can be subscribed to rather than registered. This is more generic than requiring a NetworkComms.Net delegate.

The send method was also changed to take in a connection ID rather than the connection itself, as they should not be made available outside of the Connections module.