Categorised GUI classes into subdirectories
It was getting confusing trying to navigate the GUI directory, so now it's organised by the element type.

Basic widgets are single elements drawn to the screen, typically serving only a single purpose (buttons, icons, etc.)
Compound widgets are elements that use one or more basic widgets but are not a whole composition in themselves (custom table rows, structured text fields, etc.)
Containers are widgets that contain other widgets but are not inherently visible on their own; they are usually used for layout (horizontally-centred containers, scrolling containers, etc.)
Windows and panels are final products that are drawn to the screen and consist of many of the above widget types.

There is the exception of Displayable which is a simple adaptor for the RimWorld API.