Set the height so it doesn't need to be overridden
The issue before was the spacing inadvertently included in the calculation of the flex container caused it to be drawn shorter than usual. This was fixed in 172c55186914a77c09f4698b18ff6283e2b91db2, so the workaround is no longer necessary.
Additionally, the window height has been tweaked so that the elements will always be at the right height.