Use relative time for session expiry timer
To avoid clock synchronisation issues (which seems to be a lot more common than you'd think) the timer just uses the number of milliseconds until expiry directly rather than deriving it from a timestamp.
The time may not be perfect due to latency, but we're refreshing every 15 minutes so that shouldn't be a problem.