Fix error when clearing trade item search field
When adding the entirety of an item stack to a trade, the internal list
of Things in that stack would become empty. The label calls `First()`
assuming there is always something in the stack.

Added a precondition that the stack has to have a count greater than
zero.

Also added this condition to the check which shows the 'no items
available' placeholder.

Fixes #80.